src/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
transformers==4.36.0
accelerate==0.30.1
pydantic==2.11.3
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
xformers==0.0.29.post3
ctransformers==0.2.27
llama-cpp-python==0.3.8
--extra-index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://abetlen.github.io/llama-cpp-python/cu121